[improve] Add more dependencies (#3)
* [improve] add network crate

* tmp

* [improve] modify chain_id to u64

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

* tmp

+++ api/types/src/ledger_info.rs
@@ -30,7 +30,7 @@ impl LedgerInfo {
     ) -> Self {
         let ledger_info = info.ledger_info();
         Self {
-            chain_id: chain_id.id(),
+            chain_id: chain_id.id() as u8,
             epoch: U64::from(ledger_info.epoch()),
             ledger_version: ledger_info.version().into(),
             oldest_ledger_version: oldest_ledger_version.into(),
